prediction/sentiment/requirements.txt
# Function dependencies, for example:
# package>=version
firebase_admin
google-cloud-firestore
textblob
pandas